+++ module/helpers/constants.mjs
@@ -59,14 +59,15 @@ export const ItemType = Object.freeze({
 
 export const GearType = Object.freeze({
   AFFLICTION: 'affliction',
-  AUGMENT: 'augment',
-  WEAPON: 'weapon',
   ARMOR: 'armor',
+  AUGMENT: 'augment',
+  CONTAINER: 'container',
   CORRUPTION: 'corruption',
   PET: 'pet',
-  POWER: 'power',
   POWER_ALBUM: 'power_album',
-  WARD: 'ward'
+  POWER: 'power',
+  WARD: 'ward',
+  WEAPON: 'weapon'
 });
 
 export const FrequencyType = Object.freeze({
@@ -173,6 +174,7 @@ export const DEFAULT_WEAPON_ICON = 'icons/svg/sword.svg';
 export const DEFAULT_PET_ICON = 'icons/svg/pawprint.svg';
 export const DEFAULT_AFFLICTION_ICON = 'icons/svg/stoned.svg';
 export const DEFAULT_AUGMENT_ICON = 'icons/svg/upgrade.svg';
+export const DEFAULT_CONTAINER_ICON = 'icons/svg/chest.svg';
 
 export const GEAR_ICONS = [
   DEFAULT_GEAR_ICON,
@@ -184,7 +186,8 @@ export const GEAR_ICONS = [
   DEFAULT_WEAPON_ICON,
   DEFAULT_AFFLICTION_ICON,
   DEFAULT_AUGMENT_ICON,
-  DEFAULT_PET_ICON
+  DEFAULT_PET_ICON,
+  DEFAULT_CONTAINER_ICON
 ];
 
 export const TRAIT_ICONS = [